23/05/31, 2:26am, Update 7 -
Scanner added to take user input.

Created methods to add 15-120 days or remove 15-120 days. Notes to self scattered throughout. Definitely need to clean this up a bit. Works for now though.

Gross looking if-statements to give the user the desired information.

Continuation loop added and fixed so you're not stuck in a Quantum Loop with Scott Bacula.

Time for a sandwich.